Extract query execution
